+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Library Item (Book/Magazine) description view with list of copies and buttons to perform actions on them.
 - Error handle views for 401, 404 and 500. 
 - Adding book view without test
-- Addd custom validators for evry field in add book view
+- Add custom validators for every field in add book view
 - Fix bug - Eliminate bug with already registered user registration handling
-
+- Fix validator for name and surname 
+- Fix test for registration and login